Revert "实现内存分配和内存回收"
This reverts commit 26167aa8bb01dc28a1cb50f20eb9efa0607f2d2a.

+++ src/memory/hardware/PCB.java
@@ -15,8 +15,8 @@ public class PCB {
 	private int  ir;
 	private long pid;
 	//进程在内存的位置
-	private int startPos; //start代表起点
-	private int length; //length代表这个进程的大小
+	private long startPos; //start代表起点
+	private long length; //length代表这个进程的大小
 
 	public int getId() {
 		return id;
